requirements.txt
django>=3.2,<4.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.20.0,<2.0
django-rest-swagger>=2.2.0,<2.3.0
Markdown>=3.0,<3.4
CherryPy>=18.0,<19.0
pexpect>=4.8,<4.9
python-dotenv>=0.10.3,<1.0.0
kubernetes>=11.0.0,<12.0.0
gunicorn>=20.0.0,<21.0.0
django-health-check>=3.16,<4.0
django-filter>=2.4,<3.0
